Pull all Tabs/Views related code from editor.jsx to editorTabs.jsx
